+++ identity/CHANGELOG.md
@@ -17,6 +17,9 @@
   `multiaddr::Protocol::P2p` is now type-safe and contains a `PeerId` directly, rendering this function obsolete.
   See [PR 3656].
 
+- Remove `PeerId::is_public_key` because it is unused and can be implemented externally.
+  See [PR 3656].
+
 [PR 3656]: https://github.com/libp2p/rust-libp2p/pull/3656
 [PR 3850]: https://github.com/libp2p/rust-libp2p/pull/3850
 [PR 3715]: https://github.com/libp2p/rust-libp2p/pull/3715